requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
python-multipart>=0.0.18
python-docx>=1.1.2
langchain-core>=0.3.0
langchain-openai>=0.3.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
weasyprint>=68.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.27.0
apscheduler>=3.10.0
